#6 Add API for choosing the seeking method
