+++ src/StreamSelection.cpp
@@ -1,0 +1,36 @@
+
+/*
+ *  StreamSelection.cpp
+ *  sfeMovie project
+ *
+ *  Copyright (C) 2010-2014 Lucas Soltic
+ *  lucas.soltic@orange.fr
+ *
+ *  This program is free software; you can redistribute it and/or
+ *  modify it under the terms of the GNU Lesser General Public
+ *  License as published by the Free Software Foundation; either
+ *  version 2.1 of the License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ *  Lesser General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser General Public
+ *  License along with this program; if not, write to the Free Software
+ *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+ *
+ */
+
+#include <sfeMovie/StreamSelection.hpp>
+
+namespace sfe
+{
+    StreamDescriptor StreamDescriptor::NoSelection(sfe::MediaType type)
+    {
+        StreamDescriptor descriptor;
+        descriptor.type = type;
+        descriptor.identifier = -1;
+        return descriptor;
+    }
+}
